requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
httpx
python-dotenv
sentence-transformers
faiss-cpu
docxtpl
langchain-community
requests
langchain
langgraph